+++ SAM_Systems/SAM.Analytical.Systems/Classes/Result/SystemRadiatorResult.cs
@@ -1,17 +1,19 @@
 ﻿using Newtonsoft.Json.Linq;
+using SAM.Core;
+using SAM.Core.Systems;
 using System.Collections.Generic;
 
-namespace SAM.Core.Systems
+namespace SAM.Analytical.Systems
 {
     public class SystemRadiatorResult : SystemIndexedDoublesResult, ISystemComponentResult
     {
         public SystemRadiatorResult(string uniqueId, string name, string source, IndexedDoubles heatingLoads)
-            : base(uniqueId, name, source, Query.Dictionary(RadiatorDataType.HeatingLoad, heatingLoads))
+            : base(uniqueId, name, source, Core.Systems.Query.Dictionary(RadiatorDataType.HeatingLoad, heatingLoads))
         {
         }
 
         public SystemRadiatorResult(string uniqueId, string name, string source, Dictionary<RadiatorDataType, IndexedDoubles> dictionary)
-            : base(uniqueId, name, source, Query.Dictionary(dictionary))
+            : base(uniqueId, name, source, Core.Systems.Query.Dictionary(dictionary))
         {
         }
