+++ SAM_Systems/SAM.Analytical.Systems/Classes/SystemComponent/SystemCoolingCoil.cs
