+++ SAM_Systems/SAM.Analytical.Systems/Classes/SystemComponent/SystemRadiator.cs
