+++ SAM_Systems/SAM.Analytical.Systems/Interfaces/IMechanicalSystem.cs
@@ -1,4 +1,6 @@
-﻿namespace SAM.Core.Systems
+﻿using SAM.Core.Systems;
+
+namespace SAM.Analytical.Systems
 {
     public interface IMechanicalSystem : ISystemJSAMObject
     {
