+++ SAM_Systems/SAM.Analytical.Systems/SAM.Analytical.Systems.csproj
@@ -47,11 +47,27 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Classes\HeatingGroup.cs" />
-    <Compile Include="Classes\Radiator.cs" />
-    <Compile Include="Classes\Pipe.cs" />
-    <Compile Include="Classes\Pump.cs" />
-    <Compile Include="Classes\Boiler.cs" />
+    <Compile Include="Classes\Result\SystemChilledBeamResult.cs" />
+    <Compile Include="Classes\Result\SystemDXCoilUnitResult.cs" />
+    <Compile Include="Classes\Result\SystemFanCoilUnitResult.cs" />
+    <Compile Include="Classes\Result\SystemRadiatorResult.cs" />
+    <Compile Include="Classes\Result\SystemSpaceResult.cs" />
+    <Compile Include="Classes\SystemComponent\SystemBoiler.cs" />
+    <Compile Include="Classes\SystemComponent\SystemChilledBeam.cs" />
+    <Compile Include="Classes\SystemComponent\SystemCoolingCoil.cs" />
+    <Compile Include="Classes\SystemComponent\SystemDXCoilUnit.cs" />
+    <Compile Include="Classes\SystemComponent\SystemFanCoilUnit.cs" />
+    <Compile Include="Classes\SystemComponent\SystemHeatingCoil.cs" />
+    <Compile Include="Classes\SystemComponent\SystemRadiator.cs" />
+    <Compile Include="Classes\System\AirSystem.cs" />
+    <Compile Include="Classes\System\ElectricalSystem.cs" />
+    <Compile Include="Classes\System\FluidSystem.cs" />
+    <Compile Include="Enums\ChilledBeamDataType.cs" />
+    <Compile Include="Enums\DXCoilUnitDataType.cs" />
+    <Compile Include="Enums\FanCoilUnitDataType.cs" />
+    <Compile Include="Enums\RadiatorDataType.cs" />
+    <Compile Include="Enums\SpaceDataType.cs" />
+    <Compile Include="Interfaces\IMechanicalSystem.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -65,7 +81,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Create\" />
-    <Folder Include="Enums\" />
     <Folder Include="Modify\" />
     <Folder Include="Query\" />
   </ItemGroup>
