+++ SAM_Systems/SAM.Core.Systems/Classes/SystemCircuit.cs
