+++ SAM_Systems/SAM.Core.Systems/Classes/SystemComponent.cs
