+++ SAM_Systems/SAM.Core.Systems/Classes/SystemIndexedDoublesResult.cs
